apply lessons learned to original implementation
