+++ src/applications/simple-forms/20-10207/containers/App.jsx
@@ -1,27 +1,16 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 
 import PropTypes from 'prop-types';
-import { connect, useDispatch, useSelector } from 'react-redux';
+import { connect } from 'react-redux';
 
 import { toggleValues } from 'platform/site-wide/feature-toggles/selectors';
 import FEATURE_FLAG_NAMES from 'platform/utilities/feature-toggles/featureFlagNames';
 import RoutedSavableApp from 'platform/forms/save-in-progress/RoutedSavableApp';
-import { setData } from '~/platform/forms-system/src/js/actions';
 import formConfig from '../config/form';
 import { WIP } from '../../shared/components/WIP';
 import { workInProgressContent } from '../config/constants';
 
 function App({ location, children, showForm, isLoading }) {
-  const dispatch = useDispatch();
-  const formData = useSelector(state => state.form.data);
-
-  useEffect(() => {
-    const savedForm = JSON.parse(localStorage.getItem('savedForm'));
-    if (savedForm) {
-      dispatch(setData({ ...formData, ...savedForm }));
-    }
-  }, []);
-
   if (isLoading) {
     return (
       <va-loading-indicator
